WARNING! THIS IS A MASSIVE COMMIT.
This commit is intended to perform one massive change in the code: the removal of L and T as template arguments.
Since these template arguments have pervaded (infected) the code on a very deep level, this causes many changes to the code.
Since these were sometimes the only reason for function to be templated, some template functions are now normal functions,
causing also a large .template -> .cpp moving of files, and associated changes in CMakeLists.txt files.
This causes libraries to change, some new ones are created because of that. And this in turn forces new dependencies for
other libraries. This commit compiles, everything was touched, except for the code inside Propagator.

Since this change also forced some code to be compiled for the first time, many other small bugs were also squashed.

Finally this code has begun to implement the proper structure for Weave, and the way operations shall be passd to Weave.
This was done only on a very basic level, and much is still lacking, but the structure is beginning to emerge. Some changes
in Field and Component followed this rearrangement, such as the removal of the size() member function.

Library organising is needed (it was already needed, but now the true extent is more clear). The same is true for header
file organisation.

Although we took some care in trying stuff out, some L,T template stuff is bound to have slipped between the cracks.

Secondly, we realised a need for a re-examination of the usage of memory iterators versus for-loops in many cases. This is 
now firmly on the to-do list.

Mother of all commits, could be mother of all screw-ups, lets hope not, and we can always revert... :). Simon, if you need
help in rewriting Propagator as well, I am willing to do so, it does not take long, but your code will look quite a bit
different.