Most of the code that was defined as inline, but that would probably not be inlined is moved in this commit.   Tensor, Matrix, Propagator and loadScidac for arch are affected, and quite some CMakeLists.txt files as well.
A similar thing was done for templated functions that were not template.

Some source files defined multiple functions, those have been separated as much as possible, aside from operators.

Inline and template functions almost all use the same namespace as the header file from which they are included.
This leads to a lot of spurious namespace definitions. I have moved the include directives inside the namespace
in the header file, and removed all additional namespace declarations in the inline and template files.

Because of the scale of the tests, please update as soon as possible. There is no functional change (all tests
still work for me). I hope I did not miss any file, will recheck on a clean checkout immediately after this commit.