WARNING!!!! not working copy!!!!
Added Dagger, revert etc proper to the stochastic case, modified the initialization of a stochastic propagator 
from a not stochastic one, fixed the generation of sequential meson source for stochastic case

Added a small program to inspect propagators and a program to check sequential propagator (probably the name
is not the more approrpriate one)

Problem: contraction not completely changed accordingly, the '*' between two propagator do not return the 
correct contraction due to change in the 'dagger' logic. At the moment results of contraction are wrong.

The correct way of proceeding (to be implemented in the 'get_dirac' routines) is wrote down for scalar case in 
"check_meson_3pts"